fix: resolve k8s runtime failures so smoke test passes end-to-end
- Switch all Dockerfiles from alpine to Debian-based images (maven:3.9-eclipse-temurin-17,
  eclipse-temurin:17-jre) — alpine images have no ARM64 variant; use groupadd/useradd
  instead of addgroup/adduser
- Fix JwtAuthGatewayFilter: replace DefaultServerHttpRequestBuilder.headers() lambda
  (which receives ReadOnlyHttpHeaders causing UnsupportedOperationException) with a
  ServerHttpRequestDecorator that returns a mutable HttpHeaders copy
- Add @EnableScheduling to ScoringServiceApplication and SearchServiceApplication —
  required by Spring Kafka 3.2.x @EnableKafkaRetryTopic to schedule retry tasks
- Fix controller @RequestMapping paths to match gateway routes:
  SearchConfigController /api/v1/search-configs → /api/v1/scans/search-configs
  ScoringProfileController /api/v1/scoring-profile → /api/v1/scores/scoring-profile
  JobScoreController /api/v1/job-scores → /api/v1/scores/job-scores
- Fix Flyway migration V2: score column SMALLINT → INTEGER to match JPA entity mapping
- Add @JdbcTypeCode(SqlTypes.JSON) to jsonb columns in scan-service JPA entities
  (Hibernate 6 requires explicit JSON type binding for jsonb columns)
- Fix Helm Kafka deployment: add KAFKA_PORT="" to prevent k8s service env injection crash;
  replace exec liveness probe with tcpSocket startupProbe
- Fix Helm Zookeeper deployment: replace ruok exec probe (disabled by default) with
  tcpSocket startupProbe
- Fix postgres init ConfigMap: simpler idempotent CREATE DATABASE script
- Fix smoke-test.sh: unique tenant name per run, enabled:true on search config,
  correct field name "id" for job count

+++ scoring-service/src/main/resources/db/migration/V2__create_job_scores.sql
@@ -2,7 +2,7 @@ CREATE TABLE job_scores (
     id             UUID        NOT NULL DEFAULT gen_random_uuid(),
     tenant_id      UUID        NOT NULL,
     job_listing_id UUID        NOT NULL,
-    score          SMALLINT    NOT NULL CHECK (score BETWEEN 0 AND 100),
+    score          INTEGER     NOT NULL CHECK (score BETWEEN 0 AND 100),
     reasoning      TEXT        NOT NULL,
     model          VARCHAR(100) NOT NULL,
     scored_at      TIMESTAMPTZ NOT NULL DEFAULT now(),
